Use phrase "marriage" to name components and routes
